Rename BaseServer
